refactor(_core): introduce persistence umbrella + split vectors by backend (#80, ADR 041)
Apply a uniform layout rule for multi-backend infrastructure:

- **Umbrella for same-role, different abstractions**
    _core/infrastructure/database/  → persistence/rdb/
    _core/infrastructure/dynamodb/  → persistence/nosql/dynamodb/

- **Subfolders for same abstraction, different backends**
    vectors/base_s3vector_store.py       → vectors/s3/base_store.py
    vectors/s3vector_client.py           → vectors/s3/client.py
    vectors/exceptions.py (S3-specific)  → vectors/s3/exceptions.py
    vectors/base_in_memory_vector_store  → vectors/in_memory/base_store.py
    vectors/vector_model.py               (stays at root — shared by both backends)

Key calls:
- Vectors stays top-level (role = retrieval, not CRUD persistence);
  grouping it under persistence/ would conflate two APIs that consumers
  compose (Repository + VectorStore).
- Single-implementation folders (embedding/, llm/, taskiq/, http/,
  storage/, admin/) stay flat — umbrellas / subfolders only where
  multiplicity is real.
- Domain-side infra (src/{domain}/infrastructure/database/ etc.) is
  unchanged: domains own one chosen backend, not a catalogue.
- env vars (S3VECTORS_*, DYNAMODB_*), Settings fields, and
  migrations/{dynamodb,s3vectors}/ CLI folders keep product names —
  they configure AWS products, not the generic abstraction.

Scope: 32 files touched (all domains' Repository/Model import paths,
tests, migrations/env.py, bootstraps, 6 docs). Class names are
unchanged — only module paths and three file renames.

See ADR 041 for the policy and alternatives considered.
