+++ docs/history/README.md
@@ -44,7 +44,7 @@ Each document captures the context and decision criteria for "why this technolog
 | [034](034-s3vectors-vectorstore-pattern.md) | S3 Vectors Integration with VectorStore Pattern | Accepted | Done | 2026-04-14 |
 | [035](035-embedding-service-abstraction.md) | Embedding Service Abstraction with Selector Pattern | Accepted | Done | 2026-04-14 |
 | [036](036-text-chunking-semantic-text-splitter.md) | Text Chunking with semantic-text-splitter | Accepted | Done | 2026-04-14 |
-| [041](041-vector-backends-consolidation.md) | Consolidate Vector Backends Under `_core/infrastructure/vectors/` | Accepted | Done | 2026-04-20 |
+| [041](041-vector-backends-consolidation.md) | Multi-backend Infrastructure Layout — Persistence Umbrella and Backend Subfolders | Accepted | Done | 2026-04-20 |
 
 ## Future Considerations (Open Issues)
