+++ src/_core/infrastructure/di/core_container.py
@@ -4,14 +4,16 @@ from taskiq import InMemoryBroker
 from src._core.config import settings
 from src._core.domain.value_objects.embedding_config import EmbeddingConfig
 from src._core.domain.value_objects.llm_config import LLMConfig
-from src._core.infrastructure.database.config import DatabaseConfig
-from src._core.infrastructure.database.database import Database
-from src._core.infrastructure.dynamodb.dynamodb_client import DynamoDBClient
 from src._core.infrastructure.embedding.pydantic_ai_embedding_adapter import (
     PydanticAIEmbeddingAdapter,
 )
 from src._core.infrastructure.http.http_client import HttpClient
 from src._core.infrastructure.llm.model_factory import build_llm_model
+from src._core.infrastructure.persistence.nosql.dynamodb.dynamodb_client import (
+    DynamoDBClient,
+)
+from src._core.infrastructure.persistence.rdb.config import DatabaseConfig
+from src._core.infrastructure.persistence.rdb.database import Database
 from src._core.infrastructure.storage.object_storage import ObjectStorage
 from src._core.infrastructure.storage.object_storage_client import ObjectStorageClient
 from src._core.infrastructure.taskiq.broker import (
@@ -19,7 +21,7 @@ from src._core.infrastructure.taskiq.broker import (
     create_sqs_broker,
 )
 from src._core.infrastructure.taskiq.manager import TaskiqManager
-from src._core.infrastructure.vectors.s3vector_client import S3VectorClient
+from src._core.infrastructure.vectors.s3.client import S3VectorClient
 
 
 class CoreContainer(containers.DeclarativeContainer):
