+++ src/_core/infrastructure/persistence/nosql/dynamodb/dynamodb_model.py
