+++ src/_core/infrastructure/persistence/nosql/dynamodb/exceptions.py
