+++ src/_core/infrastructure/persistence/rdb/exceptions.py
