+++ src/_core/infrastructure/vectors/s3/exceptions.py
