+++ src/docs/infrastructure/database/models/document_model.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from sqlalchemy import DateTime, Integer, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column
 
-from src._core.infrastructure.database.database import Base
+from src._core.infrastructure.persistence.rdb.database import Base
 
 
 class DocumentModel(Base):
