+++ src/docs/infrastructure/vectors/document_chunk_in_memory_store.py
@@ -5,7 +5,7 @@ from typing import Any
 from pydantic import BaseModel
 
 from src._core.domain.value_objects.rag.chunk import BaseChunkDTO
-from src._core.infrastructure.vectors.base_in_memory_vector_store import (
+from src._core.infrastructure.vectors.in_memory.base_store import (
     BaseInMemoryVectorStore,
 )
 from src._core.infrastructure.vectors.vector_model import VectorData
