+++ src/docs/infrastructure/vectors/document_chunk_s3_store.py
@@ -5,8 +5,8 @@ from typing import Any
 from pydantic import BaseModel
 
 from src._core.domain.value_objects.rag.chunk import BaseChunkDTO
-from src._core.infrastructure.vectors.base_s3vector_store import BaseS3VectorStore
-from src._core.infrastructure.vectors.s3vector_client import S3VectorClient
+from src._core.infrastructure.vectors.s3.base_store import BaseS3VectorStore
+from src._core.infrastructure.vectors.s3.client import S3VectorClient
 from src._core.infrastructure.vectors.vector_model import VectorData
 from src.docs.infrastructure.vectors.document_chunk_vector_model import (
     DocumentChunkVectorModel,
