+++ src/user/infrastructure/database/models/user_model.py
@@ -1,7 +1,7 @@
 from sqlalchemy import DateTime, Integer, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
-from src._core.infrastructure.database.database import Base
+from src._core.infrastructure.persistence.rdb.database import Base
 
 
 class UserModel(Base):
