+++ tests/integration/_core/infrastructure/persistence/nosql/dynamodb/test_dynamo_repository.py
@@ -12,16 +12,20 @@ from pydantic import BaseModel
 
 from src._core.domain.value_objects.cursor_page import CursorPage
 from src._core.domain.value_objects.dynamo_key import DynamoKey, SortKeyCondition
-from src._core.infrastructure.dynamodb.base_dynamo_repository import (
+from src._core.infrastructure.persistence.nosql.dynamodb.base_dynamo_repository import (
     BaseDynamoRepository,
 )
-from src._core.infrastructure.dynamodb.dynamodb_client import DynamoDBClient
-from src._core.infrastructure.dynamodb.dynamodb_model import (
+from src._core.infrastructure.persistence.nosql.dynamodb.dynamodb_client import (
+    DynamoDBClient,
+)
+from src._core.infrastructure.persistence.nosql.dynamodb.dynamodb_model import (
     DynamoModel,
     DynamoModelMeta,
     GSIDefinition,
 )
-from src._core.infrastructure.dynamodb.exceptions import DynamoDBNotFoundException
+from src._core.infrastructure.persistence.nosql.dynamodb.exceptions import (
+    DynamoDBNotFoundException,
+)
 
 DYNAMODB_ENDPOINT = "http://localhost:8000"
 TABLE_NAME = "test_integration_notes"
